requirements.txt
asgiref==3.8.1
Django==5.0.6
django-ckeditor-5==0.2.13
django-filter==24.2
django-jazzmin==3.0.0
djangorestframework==3.15.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.0
pillow==10.3.0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.4
redis-dict==2.5.0
setuptools==70.0.0
sqlparse==0.5.0
uritemplate==4.1.1
